+++ rust/crates/runtime/src/orchestration/spawner.rs
@@ -42,42 +42,10 @@ pub struct SpawnContext {
 
 // ─── Agent Status ───────────────────────────────────────────────────────────
 
-/// Current status of a spawned agent.
-#[derive(Debug, Clone, PartialEq)]
-pub enum AgentStatus {
-    Initializing,
-    Running { activity: String },
-    Idle,
-    Completed { result: String },
-    Failed { error: String },
-    Cancelled,
-}
-
-impl AgentStatus {
-    pub fn is_terminal(&self) -> bool {
-        matches!(
-            self,
-            Self::Completed { .. } | Self::Failed { .. } | Self::Cancelled
-        )
-    }
-}
-
-// ─── Spawned Agent Metrics ──────────────────────────────────────────────────
-
-/// Metrics tracked for a spawned agent.
-#[derive(Debug, Clone, Default)]
-pub struct SpawnedAgentMetrics {
-    pub turns_completed: u32,
-    pub tool_calls: u32,
-    pub prompt_tokens: u64,
-    pub completion_tokens: u64,
-    /// Number of permission requests sent to parent.
-    pub permission_requests: u32,
-    /// Number of permission requests approved by parent.
-    pub permission_requests_approved: u32,
-    /// Number of tools blocked by permission.
-    pub tools_blocked: u32,
-}
+// Re-export from turn-core (canonical definitions live there).
+pub use astra_turn_core::orchestration_types::{
+    AgentStatus, SpawnedAgentInfo, SpawnedAgentMetrics,
+};
 
 /// Permission summary for display purposes.
 #[derive(Debug, Clone, Default)]
@@ -113,20 +81,7 @@ pub struct SpawnedAgentState {
     pub permission_summary: PermissionSummary,
 }
 
-/// Summary info for listing agents (lighter than full state).
-#[derive(Debug, Clone)]
-pub struct SpawnedAgentInfo {
-    pub agent_id: String,
-    pub run_id: String,
-    pub parent_run_id: String,
-    pub agent_type: String,
-    pub description: String,
-    pub status: AgentStatus,
-    pub started_at: SystemTime,
-    pub metrics: SpawnedAgentMetrics,
-    /// Whether there are any permission denials.
-    pub has_permission_issues: bool,
-}
+// SpawnedAgentInfo is re-exported from orchestration_types above.
 
 impl From<&SpawnedAgentState> for SpawnedAgentInfo {
     fn from(state: &SpawnedAgentState) -> Self {
